Consolidate categories into one file
