+++ src/components/modals/CategoryModal.tsx
@@ -3,12 +3,11 @@ import * as React from 'react'
 import { TouchableHighlight, View } from 'react-native'
 import { AirshipBridge } from 'react-native-airship'
 
-import { getSubcategories, setNewSubcategory } from '../../actions/TransactionDetailsActions'
+import { Category, displayCategories, formatCategory, getSubcategories, joinCategory, setNewSubcategory, splitCategory } from '../../actions/CategoriesActions'
 import { useAsyncEffect } from '../../hooks/useAsyncEffect'
 import { useHandler } from '../../hooks/useHandler'
 import { lstrings } from '../../locales/strings'
 import { useDispatch, useSelector } from '../../types/reactRedux'
-import { Category, displayCategories, formatCategory, joinCategory, splitCategory } from '../../util/categories'
 import { scale } from '../../util/scaling'
 import { MinimalButton } from '../buttons/MinimalButton'
 import { cacheStyles, Theme, useTheme } from '../services/ThemeContext'
