+++ src/components/themed/TransactionRow.tsx
@@ -6,9 +6,9 @@ import FastImage from 'react-native-fast-image'
 import Ionicons from 'react-native-vector-icons/Ionicons'
 import { sprintf } from 'sprintf-js'
 
+import { formatCategory, splitCategory } from '../../actions/CategoriesActions'
 import { useHandler } from '../../hooks/useHandler'
 import { lstrings } from '../../locales/strings'
-import { formatCategory, splitCategory } from '../../util/categories'
 import { triggerHaptic } from '../../util/haptic'
 import { unixToLocaleDateTime } from '../../util/utils'
 import { cacheStyles, Theme, useTheme } from '../services/ThemeContext'
